+++ modules/api-svc/src/index.js
@@ -139,6 +139,63 @@ class Server extends EventEmitter {
         }
     }
 
+    _shouldUpdateInboundServer(newConf) {
+        const isInboundDifferent = !_.isEqual(this.conf.inbound, newConf.inbound);
+        const isOutboundDifferent = !_.isEqual(this.conf.outbound, newConf.outbound);
+        const isPeerJWSKeysDifferent = !_.isEqual(this.conf.peerJWSKeys, newConf.peerJWSKeys);
+        const isJwsSigningKeyDifferent = !_.isEqual(this.conf.jwsSigningKey, newConf.jwsSigningKey);
+
+        if (isInboundDifferent) {
+            this.logger.debug('Inbound config is different', {
+                oldInbound: this.conf.inbound,
+                newInbound: newConf.inbound
+            });
+        }
+        if (isOutboundDifferent) {
+            this.logger.debug('Outbound config is different (checked in inbound update)', {
+                oldOutbound: this.conf.outbound,
+                newOutbound: newConf.outbound
+            });
+        }
+
+        if (isPeerJWSKeysDifferent) {
+            this.logger.debug('Peer JWS Keys config is different', {
+                oldPeerJWSKeys: this.conf.peerJWSKeys,
+                newPeerJWSKeys: newConf.peerJWSKeys
+            });
+        }
+
+        if (isJwsSigningKeyDifferent) {
+            this.logger.debug('JWS Signing Key config is different', {
+                oldJwsSigningKey: this.conf.jwsSigningKey,
+                newJwsSigningKey: newConf.jwsSigningKey
+            });
+        }
+
+        return isInboundDifferent || isOutboundDifferent || isPeerJWSKeysDifferent || isJwsSigningKeyDifferent;
+    }
+
+    _shouldUpdateOutboundServer(newConf) {
+        const isOutboundDifferent = !_.isEqual(this.conf.outbound, newConf.outbound);
+        const isJwsSigningKeyDifferent = !_.isEqual(this.conf.jwsSigningKey, newConf.jwsSigningKey);
+
+        if (isOutboundDifferent) {
+            this.logger.debug('Outbound config is different', {
+                oldOutbound: this.conf.outbound,
+                newOutbound: newConf.outbound
+            });
+        }
+
+        if (isJwsSigningKeyDifferent) {
+            this.logger.debug('JWS Signing Key config is different', {
+                oldJwsSigningKey: this.conf.jwsSigningKey,
+                newJwsSigningKey: newConf.jwsSigningKey
+            });
+        }
+
+        return isOutboundDifferent || isJwsSigningKeyDifferent;
+    }
+
     async start() {
         await this.cache.connect();
         await this.wso2.auth.start();
@@ -224,9 +281,11 @@ class Server extends EventEmitter {
         }
 
         this.logger.isDebugEnabled && this.logger.push({ oldConf: this.conf.inbound, newConf: newConf.inbound }).debug('Inbound server configuration');
-        const updateInboundServer = !_.isEqual(this.conf.inbound, newConf.inbound)
-            || !_.isEqual(this.conf.outbound, newConf.outbound);
+        const updateInboundServer = this._shouldUpdateInboundServer(newConf);
         if (updateInboundServer) {
+            const stopStartLabel = 'InboundServer stop/start duration';
+            // eslint-disable-next-line no-console
+            console.time(stopStartLabel);
             await this.inboundServer.stop();
             this.inboundServer = new InboundServer(
                 newConf,
@@ -240,12 +299,17 @@ class Server extends EventEmitter {
                 this.emit('error', errMessage);
             });
             await this.inboundServer.start();
+            // eslint-disable-next-line no-console
+            console.timeEnd(stopStartLabel);
             restartActionsTaken.updateInboundServer = true;
         }
 
         this.logger.isDebugEnabled && this.logger.push({ oldConf: this.conf.outbound, newConf: newConf.outbound }).debug('Outbound server configuration');
-        const updateOutboundServer = !_.isEqual(this.conf.outbound, newConf.outbound);
+        const updateOutboundServer = this._shouldUpdateOutboundServer(newConf);
         if (updateOutboundServer) {
+            const stopStartLabel = 'OutboundServer stop/start duration';
+            // eslint-disable-next-line no-console
+            console.time(stopStartLabel);
             await this.outboundServer.stop();
             this.outboundServer = new OutboundServer(
                 newConf,
@@ -260,6 +324,8 @@ class Server extends EventEmitter {
                 this.emit('error', errMessage);
             });
             await this.outboundServer.start();
+            // eslint-disable-next-line no-console
+            console.timeEnd(stopStartLabel);
             restartActionsTaken.updateOutboundServer = true;
         }
 
@@ -282,13 +348,6 @@ class Server extends EventEmitter {
             await this.controlClient?.stop();
             if (this.conf.pm4mlEnabled) {
                 const RESTART_INTERVAL_MS = 10000;
-                this.controlClient = await ControlAgent.Client.Create({
-                    address: newConf.control.mgmtAPIWsUrl,
-                    port: newConf.control.mgmtAPIWsPort,
-                    logger: this.logger,
-                    appConfig: newConf,
-                });
-                this.controlClient.on(ControlAgent.EVENT.RECONFIGURE, this.restart.bind(this));
 
                 const schedulePing = () => {
                     clearTimeout(this.pingTimeout);
@@ -299,6 +358,18 @@ class Server extends EventEmitter {
                         }));
                     }, PING_INTERVAL_MS + this.conf.control.mgmtAPILatencyAssumption);
                 };
+
+                schedulePing();
+
+                this.controlClient = await ControlAgent.Client.Create({
+                    address: newConf.control.mgmtAPIWsUrl,
+                    port: newConf.control.mgmtAPIWsPort,
+                    logger: this.logger,
+                    appConfig: newConf,
+                });
+                this.controlClient.on(ControlAgent.EVENT.RECONFIGURE, this.restart.bind(this));
+
+
 
                 this.controlClient.on('ping', () => {
                     this.logger.debug('Received ping from control server');
@@ -315,7 +386,6 @@ class Server extends EventEmitter {
                     }, RESTART_INTERVAL_MS);
                 });
 
-                schedulePing();
                 restartActionsTaken.updateControlClient = true;
             }
         }